__main__ import reorganization
