+++ files/routes/__init__.py
@@ -4,6 +4,9 @@ from files.__main__ import app
 # import jinja2 helper then...
 from files.helpers.jinja2 import *
 
+# import allroutes then...
+from files.routes.allroutes import *
+
 # import routes :)
 from .admin import *
 from .comments import *
